fix: enviroment fixes for pnpm start
