+++ .eslintignore
@@ -1,0 +1,2 @@
+build
+commitlint.config.ts
