+++ .gitignore
@@ -168,3 +168,8 @@ dist
 .yarn/build-state.yml
 .yarn/install-state.gz
 .pnp.\*
+
+# Out dirs
+
+dist
+build
