+++ package.json
@@ -2,9 +2,9 @@
   "name": "pye-community",
   "version": "1.0.0",
   "description": "A community bot created for https://discord.gg/programacion (PyE server)",
-  "main": "src/index.ts",
+  "main": "build/index.js",
   "scripts": {
-    "start": "pnpm run build && node build/index.js",
+    "start": "pnpm build && node build/index.js",
     "clean": "rimraf ./build",
     "prebuild": "pnpm run clean",
     "dev": "pnpm ts-node-dev --respawn --transpile-only src/index.ts dev",
