+++ src/commands/ping.ts
@@ -1,9 +1,9 @@
-import { Client, CommandInteraction, SlashCommandBuilder } from 'discord.js';
+import { CommandInteraction, SlashCommandBuilder } from 'discord.js';
 
 export const data = new SlashCommandBuilder()
   .setName('ping')
   .setDescription('replies with pong');
 
-export async function execute(interaction: CommandInteraction, client: Client) {
+export async function execute(interaction: CommandInteraction) {
   await interaction.reply('pong!');
 }
