+++ src/index.ts
@@ -1,7 +1,6 @@
-import { Client, GatewayIntentBits, Interaction, Partials } from 'discord.js';
+import { Client, GatewayIntentBits, Partials } from 'discord.js';
 import config from './config';
 import { loadEvents } from './modules/bot/eventsLoader';
-import { loadSlashCommands } from './modules/bot/slashCommandsLoader';
 
 export const client = new Client({
   intents: [
@@ -12,15 +11,6 @@ export const client = new Client({
   ],
   partials: [Partials.Message, Partials.Channel],
   allowedMentions: { parse: ['users'] },
-});
-
-const slashCommands = loadSlashCommands();
-
-client.on('interactionCreate', async (interaction: Interaction) => {
-  if (!interaction.isChatInputCommand()) return;
-  const { commandName } = interaction;
-  const slashCommand = (await slashCommands).get(commandName);
-  slashCommand?.execute(interaction, client);
 });
 
 loadEvents(client).catch((err) => {
